@@ -1,12 +1,7 @@
-/// <reference path="../../definitions/node.d.ts"/>
-/// <reference path="../../definitions/Q.d.ts"/>
+import { EventEmitter } from 'events';
+import { ChildProcess, spawn } from 'child_process';
 
-import Q = require('q');
-import events = require('events');
-import fs = require('fs');
-import path = require('path');
-import child = require('child_process');
-let shell = require('shelljs');
+import shell from 'shelljs';
 
 function debug(message: string): void {
     if (process.env['TASK_TEST_TRACE']) {
@@ -14,45 +9,57 @@ function debug(message: string): void {
     }
 }
 
-class PSEngineRunner extends events.EventEmitter {
+// Thanks to https://stackoverflow.com/a/34637436
+class DeferredPromise {
+    public promise: Promise<void>
+    public resolve: () => void
+    public reject: (reason: any) => void
+
+    constructor() {
+        this.promise = new Promise((resolve, reject) => {
+            this.reject = reject
+            this.resolve = resolve
+        })
+    }
+}
+
+class PSEngineRunner extends EventEmitter {
     constructor() {
         super();
     }
 
-    private _childProcess: child.ChildProcess;
-    private _errors: string[];
-    private _runDeferred: Q.Deferred<void>;
+    private childProcess: ChildProcess;
+    private errors: string[];
+    private dp: DeferredPromise;
+
     public stderr: string;
     public stdout: string;
 
     public ensureKill(): void {
-        if (this._childProcess) {
-            this._childProcess.kill();
+        if (this.childProcess) {
+            this.childProcess.kill();
         }
-
-        this._childProcess = undefined;
     }
 
-    public run(psPath: string, done) {
+    public run(psPath: string, done: (err?: any) => void): void {
         this.ensureStarted();
         this.runPromise(psPath)
-        .then(() => {
-            done();
-        })
-        .fail((err) => {
-            done(err);
-        });
+            .then(() => done())
+            .catch((err) => done(err));
     }
 
     private ensureStarted(): void {
-        if (this._childProcess) {
+        if (this.childProcess) {
             return;
         }
 
+        const powershell = shell.which('powershell.exe')?.stdout;
+        if (!powershell) {
+            throw new Error('powershell.exe not found');
+        }
+
         this.emit('starting');
-        var defer = Q.defer<void>();
-        var powershell = shell.which('powershell.exe').stdout;
-        this._childProcess = child.spawn(
+        this.childProcess = spawn(
             powershell, // command
             [ // args
                 '-NoLogo',
@@ -68,51 +75,55 @@ class PSEngineRunner extends events.EventEmitter {
                 cwd: __dirname,
                 env: process.env
             });
-        this._childProcess.stdout.on(
+        this.childProcess.stdout.on(
             'data',
             (data) => {
                 // Check for special ouput indicating end of test.
                 if (('' + data).indexOf('_END_OF_TEST_ce10a77a_') >= 0) {
-                    if (this._errors.length > 0) {
-                        this._runDeferred.reject(this._errors.join('\n'));
+                    if (this.errors.length > 0) {
+                        this.dp.reject(this.errors.join('\n'));
                     } else {
-                        this._runDeferred.resolve(null);
+                        this.dp.resolve();
                     }
                 } else if (data != '\n') {
                     if (('' + data).match(/##vso\[task.logissue .*type=error/)) {
                         // The VstsTaskSdk converts error records to error issue commands.
                         debug('stdout: ' + data);
-                        this._errors.push(data);
+                        this.errors.push(data);
                     } else {
                         // Otherwise, normal stdout.
                         debug('stdout: ' + data);
                     }
                 }
             });
-        this._childProcess.stderr.on(
+        this.childProcess.stderr.on(
             'data',
             (data) => {
                 // Stderr indicates an error record was written to PowerShell's error pipeline.
                 debug('stderr: ' + data);
-                this._errors.push(data);
+                this.errors.push(data);
             });
     }
 
-    private runPromise(psPath: string): Q.Promise<void> {
+    private runPromise(psPath: string): Promise<void> {
         this.emit('running test');
-        this._errors = [];
-        this._runDeferred = Q.defer<void>();
-        this._childProcess.stdin.write(psPath + '\n')
-        return <Q.Promise<void>>this._runDeferred.promise;
+
+        this.errors = [];
+        this.dp = new DeferredPromise();
+        this.childProcess.stdin.write(psPath + '\n')
+
+        return this.dp.promise;
     }
 }
 
-let _engineRunner: PSEngineRunner = new PSEngineRunner();
+const engineRunner: PSEngineRunner = new PSEngineRunner();
 
-export function run(psPath: string, done): void {
-    _engineRunner.run(psPath, done);
+export function run(psPath: string, done: (err?: any) => void): void {
+    engineRunner.run(psPath, done);
 }
 
 export function ensureStopped(): void {
-    _engineRunner.ensureKill();
+    engineRunner.ensureKill();
 }
+
+export default { run, ensureStopped };
